Remove CostExpressionParseTreeVisitor
